# beetle-quest/api/rest/v1/openAPI.yaml
openapi: 3.0.0
info:
    title: Beetle Quest API
    version: 0.1.0
servers:
    - url: https://localhost/api/v1
...
paths:
...
    /auth/logout:
        get:
            summary: Logout
            tags: [Auth]
            security:
                - cookieAuth: []
            #     - oauth2:
            #           - user
            #           - admin
            responses:
                302:
                    description: Logout successfull.
                500:
                    description: Internal server error.
...
components:
    securitySchemes:
        cookieAuth:
            type: apiKey
            in: cookie
            name: access_token
        # oauth2:
        #     type: oauth2
        #     flows:
        #         authorizationCode:
        #             authorizationUrl: https://localhost/oauth2/authorize
        #             tokenUrl: https://localhost/oauth2/token
        #             scopes:
        #                 user: User scope
        #                 admin: Admin scope
...